# frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "dgat-backend",
        "description": "",
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "https://app.decidel.app/api",
            "description": "DGAT Backend Server"
        }
    ],
    "paths": {
...
        "/admin/gaps": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new gap (admin)",
                "description": "Creates a new gap with a specified severity and description.\nThis endpoint requires a direct severity level and does not use rule-based calculations.\n\nRequired fields:\n- `dimension_id` (UUID)\n- `gap_description` (string)\n- `gap_severity` (enum: \"LOW\", \"MEDIUM\", \"HIGH\")\n\nExample:\n{\n\"dimension_id\": \"3fa85f64-5717-4562-b3fc-2c963f66afa6\",\n\"gap_description\": \"A significant gap in the marketing dimension.\",\n\"gap_severity\": \"HIGH\"\n}",
                "operationId": "admin_create_gap",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminCreateGapRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Gap created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGapResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Dimension assessment not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/assessments/{assessment_id}/gaps": {
            "get": {
                "tags": [
                    "Gaps"
                ],
                "summary": "List gaps by assessment",
                "description": "Retrieves a paginated list of gaps for a specific assessment across all dimensions.",
                "operationId": "list_gaps_by_assessment",
                "parameters": [
                    {
                        "name": "assessment_id",
                        "in": "path",
                        "description": "Assessment ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (1-based)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gaps retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePaginatedGapResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Assessment not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/dimension-assessments/{dimension_assessment_id}/gaps": {
            "get": {
                "tags": [
                    "Gaps"
                ],
                "summary": "List gaps by dimension assessment",
                "description": "Retrieves a paginated list of gaps for a specific dimension assessment.",
                "operationId": "list_gaps_by_dimension_assessment",
                "parameters": [
                    {
                        "name": "dimension_assessment_id",
                        "in": "path",
                        "description": "Dimension Assessment ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (1-based)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gaps retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePaginatedGapResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Dimension assessment not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/gaps": {
            "get": {
                "tags": [
                    "Gaps"
                ],
                "summary": "List all gaps with pagination",
                "description": "Retrieves a paginated list of all gaps in the system.",
                "operationId": "list_gaps",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (1-based)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language code to filter content (e.g. en, fr, pt, ss). Defaults to en.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gaps retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePaginatedGapResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/gaps/{id}": {
            "get": {
                "tags": [
                    "Gaps"
                ],
                "summary": "Get a specific gap by ID",
                "description": "Retrieves the details of a specific gap using its unique identifier.",
                "operationId": "get_gap",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gap ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gap details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGapResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Gap not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Gaps"
                ],
                "summary": "Update an existing gap",
                "description": "Optionally updates gap size (and recalculates severity) and/or description.",
                "operationId": "update_gap",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gap ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGapRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Gap updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGapResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Gap not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gaps"
                ],
                "summary": "Delete a gap",
                "operationId": "delete_gap",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gap ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gap deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEmpty"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Gap not found"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminCreateGapRequest": {
                "type": "object",
                "required": [
                    "dimension_key",
                    "gap_description",
                    "gap_severity"
                ],
                "properties": {
                    "dimension_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Language-specific dimension ID (resolved by backend from dimension_key + language)",
                        "nullable": true
                    },
                    "dimension_key": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Stable cross-language dimension identifier",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "gap_description": {
                        "type": "string",
                        "description": "explicit description to override",
                        "example": "Significant gap"
                    },
                    "gap_severity": {
                        "$ref": "#/components/schemas/GapSeverity"
                    },
                    "language": {
                        "type": "string",
                        "description": "Language code for this content entry (e.g. en, fr, pt, ss)",
                        "default": "en"
                    }
                },
                "example": {
                    "dimension_key": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                    "gap_description": "A significant gap in the marketing dimension.",
                    "gap_severity": "HIGH",
                    "language": "en"
                }
            },
...
            "ApiResponseGapResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "type": "object",
                                "required": [
                                    "gap_id",
                                    "dimension_id",
                                    "gap_size",
                                    "gap_severity",
                                    "calculated_at",
                                    "created_at",
                                    "updated_at"
                                ],
                                "properties": {
                                    "calculated_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "When the gap was calculated",
                                        "example": "2023-01-01T00:00:00Z"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "When the gap was created",
                                        "example": "2023-01-01T00:00:00Z"
                                    },
                                    "dimension_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "description": "Reference to the dimension",
                                        "example": "550e8400-e29b-41d4-a716-446655440002"
                                    },
                                    "gap_description": {
                                        "type": "string",
                                        "description": "Optional description of the gap",
                                        "example": "Significant gap in digital transformation",
                                        "nullable": true
                                    },
                                    "gap_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "description": "Unique identifier for the gap",
                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                    },
                                    "gap_severity": {
                                        "$ref": "#/components/schemas/GapSeverity"
                                    },
                                    "gap_size": {
                                        "type": "integer",
                                        "format": "int32",
                                        "description": "Numeric size of the gap",
                                        "example": 3
                                    },
                                    "updated_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "description": "When the gap was last updated",
                                        "example": "2023-01-01T00:00:00Z"
                                    }
                                },
                                "example": {
                                    "gap_id": "550e8400-e29b-41d4-a716-446655440000",
                                    "dimension_id": "550e8400-e29b-41d4-a716-446655440002",
                                    "gap_size": 3,
                                    "gap_severity": "HIGH",
                                    "gap_description": "Significant gap in digital transformation",
                                    "calculated_at": "2023-01-01T00:00:00Z",
                                    "created_at": "2023-01-01T00:00:00Z",
                                    "updated_at": "2023-01-01T00:00:00Z"
                                }
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "GapSeverity": {
                "type": "string",
                "enum": [
                    "LOW",
                    "MEDIUM",
                    "HIGH"
                ],
                "example": "HIGH"
            },
...
            "UpdateGapRequest": {
                "type": "object",
                "description": "Update gap request",
                "properties": {
                    "gap_description": {
                        "type": "string",
                        "description": "New human-readable description",
                        "nullable": true
                    },
                    "gap_severity": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/GapSeverity"
                            }
                        ],
                        "nullable": true
                    },
                    "gap_size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "New gap size; if provided, severity will be recalculated",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "description": "Language code for this content entry (e.g. en, fr, pt, ss)",
                        "nullable": true
                    }
                }
            },
...
        }
    },
...
}